Debugging SW_ProcessingElement (1)
Tested ScoringModule against ssearch36/fasta36 and
swalign-0.3.3(opensource python module) with data1.fa. There are errors
only during calculation of the scores for sequences with gaps(In-dels)
this suggests that there might be a bug in the "I" matrix logic. (db3,
db7, db8, db11, db12 & db16 in data1.fa)

- Added SW_ProcessingElement_V0.1.v: added enable control for the
combinational logic.
- result stays valid at the output until the next new sequence is
streamed in.
- small tweaks in the testbenches.